Moving ScriptModule:: methods used for "#UseXXX" to separate file.
Adding ScriptModulesAddNames.h

Reason, for easier overview.

Also some minor rearrangements and comment adjustments.
