+++ AutoHotkeyx.vcxproj
@@ -259,6 +259,7 @@
     <ClInclude Include="source\resources\resource.h" />
     <ClInclude Include="source\script.h" />
     <ClInclude Include="source\ScriptModules.h" />
+    <ClInclude Include="source\ScriptModulesAddNames.h" />
     <ClInclude Include="source\ScriptModulesDefines.h" />
     <ClInclude Include="source\script_com.h" />
     <ClInclude Include="source\script_func_impl.h" />
